Renamed GirafRest.Shared to GirafRest.GirafShared
